Update list.m3u and icons
We lost some good television today due to the weird c2eds IP thing suddenly deciding to not work. Overload? Busted by the cops? Just forget it. I removed the religious freaks and garbage tabloid channels too (Kurir/Red in sec.1, InfoWars/Newsmax in sec.2) as nobody has ever willingly watched this garbage in the entire history of the universe. HOWEVER! We gained a good channel today! (2 actually). First up is Parlamentarni kanal, which is a Montenegrin state-owned channel (RTCG) that broadcasts sessions of our Parliament/Government/judiciary/some other corrupt institution. Second is TV5Monde Info, which idk I don't speak French but it looks decent and is in the EBU.

(I also replaced the RTCG links. I think 2 are geo-blocked, but SAT is made for diaspora and practically nobody outside of Montenegro/the diaspora watches it. After all, I don't want the police to knock on my door for using some weird Serbian SMIL site. This is practically the same thing, but an official link.)

Thank you (nobody. nobody uses this except me) for using this list.